Add bash script to run models collectively
This script worked without internally activating the conda environment
when run on the command line and not as a job. When I run it as a job, it wasn't
inside the conda env when it was running things from the script. I fixed that
by adding the eval and conda activate lines, but when it tries to format the
datasets, it still behaves as if the enviornment is not activated, saying that
it can't find en_core_web_sm. I'm very unclear what's going on here, coming back
to it tomorrow.